Fixes #26: Change order to Runtime->Mission->Deployment Type
