+++ src/main/java/io/openshift/launchpad/ui/booster/ChooseRuntimeStep.java
@@ -8,7 +8,6 @@
 package io.openshift.launchpad.ui.booster;
 
 import java.util.Iterator;
-import java.util.Optional;
 
 import javax.inject.Inject;
 
@@ -16,7 +15,6 @@ import org.jboss.forge.addon.ui.context.UIBuilder;
 import org.jboss.forge.addon.ui.context.UIContext;
 import org.jboss.forge.addon.ui.context.UIExecutionContext;
 import org.jboss.forge.addon.ui.context.UINavigationContext;
-import org.jboss.forge.addon.ui.context.UIValidationContext;
 import org.jboss.forge.addon.ui.input.UISelectOne;
 import org.jboss.forge.addon.ui.metadata.UICommandMetadata;
 import org.jboss.forge.addon.ui.metadata.WithAttributes;
@@ -27,9 +25,7 @@ import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
 import org.jboss.forge.addon.ui.wizard.UIWizardStep;
 
-import io.openshift.launchpad.catalog.Booster;
 import io.openshift.launchpad.catalog.BoosterCatalogService;
-import io.openshift.launchpad.catalog.Mission;
 import io.openshift.launchpad.catalog.Runtime;
 
 /**
@@ -58,10 +54,7 @@ public class ChooseRuntimeStep implements UIWizardStep
          runtime.setItemLabelConverter(Runtime::getId);
       }
 
-      runtime.setValueChoices(() -> {
-         Mission mission = (Mission) context.getAttributeMap().get(Mission.class);
-         return catalogService.getRuntimes(mission);
-      });
+      runtime.setValueChoices(catalogService.getRuntimes());
 
       runtime.setDefaultValue(() -> {
          Iterator<Runtime> iterator = runtime.getValueChoices().iterator();
@@ -69,20 +62,6 @@ public class ChooseRuntimeStep implements UIWizardStep
       });
 
       builder.add(runtime);
-   }
-
-   @Override
-   public void validate(UIValidationContext context)
-   {
-      UIContext uiContext = context.getUIContext();
-      Mission mission = (Mission) uiContext.getAttributeMap().get(Mission.class);
-
-      Optional<Booster> booster = catalogService.getBooster(mission, runtime.getValue());
-      if (!booster.isPresent())
-      {
-         context.addValidationError(runtime,
-                  "No booster found for mission '" + mission + "' and runtime '" + runtime.getValue() + "'");
-      }
    }
 
    @Override
@@ -97,7 +76,7 @@ public class ChooseRuntimeStep implements UIWizardStep
    public NavigationResult next(UINavigationContext context) throws Exception
    {
       context.getUIContext().getAttributeMap().put(Runtime.class, runtime.getValue());
-      return Results.navigateTo(MetadataStep.class);
+      return Results.navigateTo(ChooseMissionStep.class);
    }
 
    @Override
