+++ src/main/java/io/openshift/launchpad/ui/booster/NewProjectWizard.java
@@ -44,7 +44,7 @@ public class NewProjectWizard implements UIWizard
    @Override
    public NavigationResult next(UINavigationContext context) throws Exception
    {
-      return Results.navigateTo(ChooseDeploymentTypeStep.class);
+      return Results.navigateTo(ChooseRuntimeStep.class);
    }
 
    @Override
